% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial clean fcns.R
\name{absolute.validate}
\alias{absolute.validate}
\title{absolute.validate}
\usage{
absolute.validate(x)
}
\description{
This cycles through all the "weird tricks" I've found to ensure that
geometries are not only valid, but remain so after st_intersection or other
manipulations. Buffer 0 trick is useful for these case but can alter
geometry: https://github.com/r-spatial/sf/issues/347
}
