% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-fcns.R
\name{buffered.hull}
\alias{buffered.hull}
\title{buffered.hull}
\usage{
buffered.hull(x, buffer = 300)
}
\description{
Creates a buffered convex hull around a polygon. Used to filter proximate
highways given a place.
}
