% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-fcns.R
\name{build_identifier_df}
\alias{build_identifier_df}
\title{build_identifier_df}
\usage{
build_identifier_df(df, split_cols)
}
\description{
Helper function for fcns that split list of division types x regions. Designed for retaining
identifiers; i.e., cz & cz_name for regions.
}
