% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-fcns.R
\name{conic.transform}
\alias{conic.transform}
\title{conic.transform}
\usage{
conic.transform(stdf, crs = "+proj=lcc +lon_0=-90 +lat_1=33 +lat_2=45")
}
\description{
Transforms sf spatial data to a conic projection with distance in meters
}
