% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct.pops}
\alias{ct.pops}
\title{census tract populations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74133 rows and 4 columns.
}
\usage{
ct.pops
}
\description{
Contains tract identifiers, and some xwalk information and population by
tract. Used internally to filter polygons by interpolated population when
necessary. Can use helper fcn to attach geometry.
}
\keyword{datasets}
