% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db or api helper fcns.R
\name{db.query.and.cutout.water}
\alias{db.query.and.cutout.water}
\title{db.query.and.cutout.water}
\usage{
db.query.and.cutout.water(
  con,
  region,
  sf.simplify = TRUE,
  trim.unnamed = T,
  trim.by.area = 2e+06
)
}
\arguments{
\item{con}{Active connection with a divs.water_tigris table}

\item{region}{an sf object from which to erase water areas}

\item{sf.simplify}{Whether to apply rmapshaper::ms_simplify to water areas before
trimming}

\item{trim.unnamed}{whether or not to retain unnamed water areas. These tend to be
very small and may introduce a lot of unneeded complexity to supplied region.}
}
\description{
Queries a spatial database for tigris water areas and erases those water areas
from the supplied region to return only land areas. I could refactor this
connection to just get water areas from \code{tigris} too...
}
