% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial clean fcns.R
\name{denode.lines}
\alias{denode.lines}
\title{denode.lines}
\usage{
denode.lines(x, group.cols)
}
\arguments{
\item{x}{an \code{sf} multi/linestring object to collapse}

\item{group.cols}{grouping columns to merge /union by.}
}
\description{
Dissolves constitutive line segments into longer lines wherever possible.
Specify in \code{group.cols} which columns should be kept in output. Workflow
of function is: Union -> merge -> explode -> add ids.
}
