% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta polys wrapper fcn.R
\name{dissolve.to.boundaries.w.ID}
\alias{dissolve.to.boundaries.w.ID}
\title{dissolve.to.boundaries.w.ID}
\usage{
dissolve.to.boundaries.w.ID(sf, divtype, admin)
}
\arguments{
\item{sf}{sf object}

\item{divtype, admin}{additional identifiers to be bundled with the dissolved sf.}
}
\description{
Dissolves polygons to linear divisions, keeping some identifier columns specified
with dtype and admin aguments. Drops other pre-existing columns
}
