% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-fcns.R
\name{ez.explode}
\alias{ez.explode}
\title{ez.explode}
\usage{
ez.explode(x)
}
\description{
Drops units from any columns that have them and then calls
rmapshaper::ms_explode(), which cannot handle units.
}
