% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys fcns.R, R/spatial clean fcns.R
\name{fill.gaps}
\alias{fill.gaps}
\title{fill.hwy.gaps}
\usage{
fill.gaps(hwy, threshold = 200, return.gap.map = F, ...)

fill.gaps(hwy, threshold = 200, return.gap.map = F, ...)
}
\description{
maps fill.single.gap across output of previous functions. Takes ~denoded~
lines; expects nhpn hwy data.

maps fill.single.gap across output of previous functions. Takes ~denoded~
lines; expects nhpn hwy data.
}
