% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys fcns.R, R/spatial clean fcns.R
\name{fill.single.gap}
\alias{fill.single.gap}
\title{fill.single.gap}
\usage{
fill.single.gap(edge, nodes, threshold = 50, verbose = T, ...)

fill.single.gap(edge, nodes, threshold = 50, verbose = T, ...)
}
\arguments{
\item{edge}{Single line segment-- 1 row of larger sf \code{lines}}

\item{threshold}{Threshold in crs units. Fill gap if distance between endpoint of
given edge and another in \code{lines} is < this threshold.}

\item{verbose}{Whether to print alert when a gap is filled.}

\item{lines}{Other lines to check with \code{edge} to possibly fill gap to edge
to.}
}
\description{
Uses segment endpoints and connects those within threshold distance. Creates
continuous line that I think is as appropriate as possible.

Uses segment endpoints and connects those within threshold distance. Creates
continuous line that I think is as appropriate as possible.
}
