% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{filter.colinear.nodes}
\alias{filter.colinear.nodes}
\title{filter.colinear.nodes
Runs if \code{filter.colinear.node.threshold} is set to not NULL, in which
case it removes ray nodes in given threshold proximity of one another. In
general, this should be left as NULL when using NHPN data.}
\usage{
\method{filter}{colinear.nodes}(
  place,
  potential.ray.nodes,
  filter.colinear.node.threshold = NULL,
  verbose = F,
  ...
)
}
\description{
filter.colinear.nodes
Runs if \code{filter.colinear.node.threshold} is set to not NULL, in which
case it removes ray nodes in given threshold proximity of one another. In
general, this should be left as NULL when using NHPN data.
}
