% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{filter.ineligible.ray.nodes}
\alias{filter.ineligible.ray.nodes}
\title{filter.ineligible.ray.nodes}
\usage{
\method{filter}{ineligible.ray.nodes}(
  place,
  potential.ray.nodes,
  ray.node.distance.threshold = 100,
  verbose = F,
  ...
)
}
\description{
Sometimes there is a break or gap in a hwy route just outside of a city,
causing direct endpoint analysis to count extra rays. This filters out ray
nodes if they belong to the same highway route and are within a distance
threshold of each other. #(realizing i probably made this redundant w/ the filtering segments by length floor)
}
