% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial clean fcns.R
\name{find.endpoint.nodes}
\alias{find.endpoint.nodes}
\title{find.endpoint.nodes}
\usage{
find.endpoint.nodes(x, line.ids = c("signt1", "sign1"))
}
\description{
Gets all endpoints of all lines of a LINESTRING sf object. Returns an sf object
with points representing nodes, with a column "n" representing the number of
line segments that start or end with that node.
Requires an "id" column in segment sf.
adopts workflow from https://www.r-spatial.org/r/2019/09/26/spatial-networks.html
}
