% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{Get.bundled.ray.output}
\alias{Get.bundled.ray.output}
\title{Get.bundled.ray.output}
\usage{
Get.bundled.ray.output(place, hwy, include.map = TRUE, ...)
}
\arguments{
\item{include.map}{whether or not to generate a map along with the measure.
Defaults to TRUE; set to FALSE to save time.}

\item{...}{additional parameters passed onto mapview, if map is being returned.}
}
\description{
Calls \code{hwys2ray.nodes}, which finds all line start/endpoints that would
constitute a ray. This function counts those up and bundles with a map of rays if
\code{include.map} is TRUE
}
