% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db or api helper fcns.R
\name{get.clean.tigris.water}
\alias{get.clean.tigris.water}
\title{get.clean.tigris.water}
\usage{
get.clean.tigris.water(
  czs = NULL,
  counties = NULL,
  trim.unnamed = T,
  out.crs = NULL,
  ...
)
}
\arguments{
\item{trim.unnamed}{whether or not to retain unnamed water areas. These tend to be
very small and may introduce a lot of unneeded complexity to supplied region.}

\item{out.crs}{output crs. Can be left as NULL to retain default from tigris.}

\item{...}{Additional arguments passed on to \code{tigris::area_water}}

\item{cz, counties}{a character vector containing CZ or county ids. County ids
should be 5 characters (concatenated state + county fp codes). One of these
should be supplied, the other left as null.}
}
\description{
Queries a spatial database for tigris water areas and erases those water areas
from the supplied region to return only land areas. I could refactor this
connection to just get water areas from \code{tigris} too...
}
