% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db or api helper fcns.R
\name{get.water.for.region}
\alias{get.water.for.region}
\title{get.water.for.region}
\usage{
get.water.for.region(..., size.min = 5e+06)
}
\arguments{
\item{...}{CZ, CBSA, county, or places, passed on to \code{xwalks::x2cos}}

\item{size.min}{Minimum size in m^2, after internal boundaries are resolved (if a
water area is represented by multiple contiguous polygons)}
}
\value{
water areas for region.
}
\description{
Given a cz,cbsa, or county identifier/geoid, wraps \code{tigris::area_water} for the
overlapping counties, interfacing with my \code{xwalks} library to get overlap.
Returns all water areas above the size minimum for the given region.
}
