% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div proximity fcns.R
\name{get_azimuthal_projection_centroid}
\alias{get_azimuthal_projection_centroid}
\title{get_azimuthal_projection_centroid}
\usage{
get_azimuthal_projection_centroid(x)
}
\description{
Gets a centroid of sf object x, or midpoint if a straight line is provided
}
