% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys fcns.R
\name{handle.overlaps}
\alias{handle.overlaps}
\title{handle.overlaps}
\usage{
handle.overlaps(x)
}
\description{
Some shapes (loops, nested loops) get a series of overlapping
polygons from st_split. Example, minneapolis BTS rails. This cuts out areas
so all sub polygons are non-overlapping.
}
