% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{hwys2ray.nodes}
\alias{hwys2ray.nodes}
\title{hwys2ray.nodes}
\usage{
hwys2ray.nodes(
  boundary,
  trimmed.hwys,
  min.segment.length = 10,
  verbose = T,
  ...
)
}
\description{
From the branch in above fcn where it checks hwy types to include, this
generates ray-constituting nodes. Can filter out very small segments; by
default removes those w/ less than .5km in place boundary
}
