% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{initial.hwy2ray.subset}
\alias{initial.hwy2ray.subset}
\title{initial.hwy2ray.subset}
\usage{
initial.hwy2ray.subset(
  place,
  hwy.sf,
  hwy.types = NULL,
  buffer.meters = 300,
  ...
)
}
\description{
Sets up road objects for ray measures. Takes intersection of hwys with the
place's buffered convex hull, trims NA routes and aggregates by type, trims to length
floor, so each route must have a minimum distance within place boundary.
}
