% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lac_codes}
\alias{lac_codes}
\title{limited-access hwy codes}
\format{
An object of class \code{numeric} of length 4.
}
\usage{
lac_codes
}
\description{
Identifiers in the FCLASS column of the NHPN highway dataset associated with
limited-access highways.
}
\keyword{datasets}
