% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{largest.centers.in.region}
\alias{largest.centers.in.region}
\title{largest.centers.in.region}
\usage{
largest.centers.in.region(centers, region)
}
\description{
Takes population centers, represented as a shapefile (probably of places or
sub-county divisions), with a column for population. For each larger region
in the second argument, finds the largest population center of that larger area.
Note do not include duplicate columns in the two supplied dataframes.
Returns the largest population center in each larger region as a dataframe with point geometries
}
