% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{largest.plc.in.cz}
\alias{largest.plc.in.cz}
\title{largest place in commuting zone}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 740 rows and 4 columns.
}
\usage{
largest.plc.in.cz
}
\description{
Used for generating rays.
}
\keyword{datasets}
