% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div proximity fcns.R
\name{mean_centroid_dist2div}
\alias{mean_centroid_dist2div}
\title{get_mean_centroid_dist2div}
\usage{
mean_centroid_dist2div(
  centroid_df,
  div_df,
  identifier_column,
  pop_col = "population",
  verbose = FALSE
)
}
\description{
Calculated mean centroid distance within a study area to geographic
divisions. Takes as input a spatial dataframe for centroids and another with
divisions with a shared geographic identifier column.
}
