% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div proximity fcns.R
\name{mean_centroid_dist2div.region_mapper}
\alias{mean_centroid_dist2div.region_mapper}
\title{mean_centroid_dist2div.region_mapper}
\usage{
mean_centroid_dist2div.region_mapper(
  centroid_df,
  div_df,
  geo_level,
  verbose = F
)
}
\description{
Given a df of centroids and a df of divisions, with common region (cz, cbsa,
county) identifiers, this maps the \code{mean_centroid_dist2div} function
across the identified regions. \code{geo_level} parameter should identify the region.type.
}
