% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys fcns.R
\name{Polys.wrapper}
\alias{Polys.wrapper}
\title{Polys.wrapper}
\usage{
Polys.wrapper(region, div.sf, fill.gaps = F, ...)
}
\arguments{
\item{region}{1 sf polygon to create sub-division measures for}

\item{div.sf}{sf object with division.}

\item{fill.gaps}{Whether to check for and may fill small gaps between line
segments of division. Specify \code{threshold} for maximum gap size to
be filled in.}

\item{threshold}{maximum gap size to be filled in, in meters, if fill.gaps is
TRUE.}

\item{div.identifier.column}{If 'div.sf' has an identifier column to be used
to filter by, supply here. Must be supplied for subsequent arguments to
work. For NHPN data, this is likely "SIGNT1"}

\item{divs.always.include}{Always include divisions that have this value in
the identifier column. I.e., Interstate for hwy identifier.}

\item{divs.include.intersecting}{Include additional divisions that don't the
have above identifier, but which intersect with divs that do.}

\item{remove.NA.divs}{Whether or not to remove all divisions with an NA in
the identifier column.}

\item{negative.buffer}{Shrink region by this amount only for calculating
polygons. Useful for handling shpfiles w/ different resolutions, especially
for regions along an international border. Defaults to 100m.}

\item{min.size}{Minimum (area) size for population, in meters. Defaults to
5e5, or 1/2 a km^2. Set to NULL to not filter by size.}

\item{min.population.count}{Minimum population that must fall into a polygon
subdivision for it to be counted. Population is interpolated from census
tracts based on \% area overlap of CT within sub-polygon. Set to NULL or 0
to filter by percent instead.}

\item{min.population.perc}{Minimum percent \% of population that must fall
into a polygon subdivision for it to be counted. Population is interpolated
from census tracts based on \% area overlap of CT within sub-polygon.}

\item{return.sf}{If true, returns an sf object, w/ one row per polygon
subdivision, that can easily be mapped. If false (default), returns a one
row df for region with columns for region identifiers and number of
polygons.}
}
\description{
Wraps all the functions used to generate the polygonal subdivision measure.
}
\section{Passed to Fix.all.hwys fcn}{
NA
}

\section{Passed to subsetting fcn subset.polys.divs}{
NA
}

\section{Passed to polygonal.div}{
NA
}

