% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{prepped.hwys.to.rays}
\alias{prepped.hwys.to.rays}
\title{prepped.hwys.to.rays}
\usage{
prepped.hwys.to.rays(
  place,
  hwy,
  always.include = c("I"),
  include.intersecting = FALSE,
  include.map = TRUE,
  ...
)
}
\description{
Does most of the work to generate rays. Wrapped by count.rays but it's kept
separate so it can be called more quickly with just a place boundary and
prepped hwy sf. Uses just hwy types included in "always.include" or also
includes add'l hwys that intersect those core hwy types.
}
