% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad-conservation-fcns.R
\name{protected.area.by.type.by.tract}
\alias{protected.area.by.type.by.tract}
\title{protected.area.by.type.by.tract}
\usage{
protected.area.by.type.by.tract(pad, cts, category.colm = NULL)
}
\arguments{
\item{pad}{data for protected areas}

\item{cts}{data for census tracts. Will assume geoid column}

\item{category.colm}{string for the category by which to get protected area
purpose}
}
\description{
protected.area.by.type.by.tract
}
