% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div proximity fcns.R
\name{proximity_div_proportion}
\alias{proximity_div_proportion}
\title{proximity_div_proportion}
\usage{
proximity_div_proportion(
  centroid_df,
  div_df,
  identifier_column,
  k = 400,
  pop_col = "population",
  verbose = FALSE
)
}
\description{
Calculates the proportion of a population in a study area k meters from division.
pop_col specifies the colname for column in \code{centroid_df} that contains CT populations
#' @export
}
