% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div proximity fcns.R
\name{proximity_div_proportion.region_mapper}
\alias{proximity_div_proportion.region_mapper}
\title{proximity_div_proportion.region_mapper}
\usage{
proximity_div_proportion.region_mapper(
  centroid_df,
  div_df,
  geo_level,
  k = 800,
  verbose = F
)
}
\description{
Given a df of centroids and a df of divisions, with common region (cz, cbsa,
county) identifiers, this maps the \code{proximity_div_proportion} function
across the identified regions. \code{geo_level} parameter should identify the region.type.
}
