% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-fcns.R
\name{region.reorg}
\alias{region.reorg}
\title{region.reorg}
\usage{
region.reorg(df, region.type)
}
\arguments{
\item{df}{table to rename}

\item{region.type}{A column in df containing region identifiers, with column
itself indicating level or type of region.   #' not-exported region.reorg}
}
\description{
Turns columns with i.e., cz & cz_name to region.id/region.type format that is
expected in a variety of places in code that I've been writing.
}
\details{
i may delete here; similar fcn in geoseg
}
