% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys fcns.R
\name{subset.polys.divs}
\alias{subset.polys.divs}
\title{subset.polys.divs}
\usage{
\method{subset}{polys.divs}(
  region,
  div.sf,
  div.identifier.column = NULL,
  always.include = NULL,
  include.intersecting = F,
  remove.NA.divs = T,
  ...
)
}
\arguments{
\item{region}{1-row sf polygon with \code{region.id}/\code{region.type} columns to create
sub-division measures for}

\item{div.sf}{sf object with division.}

\item{div.identifier.column}{If 'div.sf' has an identifier column to be used to
filter by, supply here. Must be supplied for other subset arguments to work;
\code{NULL} by default.}

\item{always.include}{Always include divisions that have this value in the
identifier column}

\item{include.intersecting}{Include additional divisions that don't have above
identifier, but which intersect with divs that do.}

\item{remove.NA.divs}{Whether or not to remove all divisions with an NA in the
identifier column.}
}
\description{
Preps "division" data, whether water, highways, etc, to generate polygon measure
for supplied region. If, as with NHPN data, the divisions have an "identifier
column," the later arguments can be specified to control what divisions are kept
for polygon creation.
}
