% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract-lvl-divM-fcns.R
\name{tracts.across.water}
\alias{tracts.across.water}
\title{tracts.across.water}
\usage{
tracts.across.water(
  cz = NULL,
  cbsa = NULL,
  ctsf = NULL,
  .cos = NULL,
  area.floor = 0,
  write.path = NULL,
  ...
)
}
\arguments{
\item{.cos}{counties, as downloaded from tigris. Downloads them if left as
null.}

\item{area.floor}{area floor for water polygon in sq km. I think leaving at 0
and effectively filtering with the tract overlay will be best.}

\item{write.path}{a file path to write output to.}
}
\description{
Water divisions are areas, rather than
}
