% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract-lvl-divM-fcns.R
\name{Wrapper_gen.tract.within.distance}
\alias{Wrapper_gen.tract.within.distance}
\title{Wrapper_gen.tract.within.distance}
\usage{
Wrapper_gen.tract.within.distance(
  cz = NULL,
  cbsa = NULL,
  divs,
  cutout.water = F,
  clean.nhpn = F,
  ...
)
}
\description{
Wrapper_gen.tract.within.distance
}
