% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad-conservation-fcns.R
\name{Wrapper_pad.area.by.nbhd}
\alias{Wrapper_pad.area.by.nbhd}
\title{Wrapper_pad.area.by.nbhd}
\usage{
Wrapper_pad.area.by.nbhd(
  statefp,
  pad.dir,
  category.colms = c("featclass", "own_type", "own_name", "mang_type", "mang_name",
    "des_tp", "gap_sts"),
  geo.query = tigris::tracts,
  geo.yr = 2021,
  simplify.geos = F
)
}
\arguments{
\item{statefp}{FIPS code for state to generate}

\item{pad.dir}{Directory of USGS PAD geodatabase.}

\item{category.colms}{Column names in PAD data by which we want protected
area measures disaggregated.}

\item{geo.query}{\code{tigris} function to get tracts of block groups.}

\item{geo.yr}{Year of geographies we want.}

\item{simplify.geos}{T/F whether to run st_simplify on PAD data before
measure generation.}
}
\description{
Wraps all steps to generate measures for protected areas by type (and
overall) for a given state, by tract or block group.
}
