% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{count.rays}
\alias{count.rays}
\title{count.rays}
\usage{
count.rays(place.geoid, place.sf, hwy.sf, remove.holes = FALSE, ...)
}
\arguments{
\item{remove.holes}{Remove holes from places before counting rays. If a hwy
starts/ends in a hole, it will be counted as a ray unless this is set to
TRUE}

\item{always.include}{Type(s) of highways to always include for ray measure.
By default only interstates}

\item{include.intersecting}{Whether or not to include additional highway
types if they intersect with one or more of the core/always included hwy
types.}

\item{hwy.types}{Other types of hwys to include if and only if they intersect
with the core, "always.included" type(s). Includes all non-NA routes if
NULL (default). Only relevant if \code{include.intersecting} is true.}

\item{buffer.meters}{Amount of padding around the Place to retain when
trimming highways to place and surrounding area. Defaults to 300 meters.}

\item{length.floor}{Filters highways by route if the total length of all
portions of the highway within the area is below this threshold. Default is
1 km.}

\item{min.segment.length}{Filters highway segments by length. I.e., if a
highway route branches into multiple segments, this will filter segments
that have less than this threshold within the area. Default is 500m.}

\item{ray.node.distance.threshold}{Sometimes there is a small gap in a
highway at an interchange or something. If this happens outside of a city,
we should make sure both sides of the gap aren't counted as extra rays.
This defines the minimum distance between segment endpoints that, when they
belong to the some route, would mean that neither is counted as a ray.
Defaults to 100m. If it's NULL, skips this step.}

\item{filter.colinear.node.threshold}{if this is not NULL (default), remove
ray nodes in given threshold proximity of one another. May be useful if two
highways are colinear or parallel and nearby and you don't want to count
multiple rays.}

\item{include.map}{whether or not to generate a map along with the measure.
Defaults to TRUE; set to FALSE to save time.}

\item{verbose}{Display additional text output in console. Makes explicit some
parameters that are passed to wrapped fcns and will say where ineligible
rays are removed due to issue described above.}
}
\description{
given the GEOID of a place and sf objects containing hwy information and
place information, this calls each step to count the rays. Returns the number
of rays for the place and a map of them. Arguments are passed to wrapped
fcns. Additional arguments can also be passed to mapview::mapview to change appearance
of output map.
}
