% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial clean fcns.R
\name{fill.single.gap}
\alias{fill.single.gap}
\title{fill.single.gap}
\usage{
\method{fill}{single.gap}(edge, nodes, threshold = 200)
}
\arguments{
\item{threshold}{Threshold in crs units. Fill gap if distance between endpoint of
given edge and another in \code{lines} is < this threshold.}
}
\description{
Uses segment endpoints and connects those within threshold distance. Creates
continuous line that I think is more appropriate. First two arguments
are taken from call within fill.gaps.
}
