% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys fcns.R, R/spatial clean fcns.R
\name{fill.single.gap}
\alias{fill.single.gap}
\title{fill.single.gap}
\usage{
fill.single.gap(edge, nodes, threshold = 50, verbose = T, ...)

fill.single.gap(edge, nodes, threshold = 50, verbose = T, ...)
}
\description{
Uses segment endpoints and connects those within threshold distance. Creates
continuous line that I think is as appropriate as possible.

Uses segment endpoints and connects those within threshold distance. Creates
continuous line that I think is as appropriate as possible.
}
