% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial clean fcns.R
\name{Fix.all.hwys}
\alias{Fix.all.hwys}
\title{Fix all hwys}
\usage{
Fix.all.hwys(hwys, threshold = 200, return.gap.map = F, ...)
}
\arguments{
\item{return.gap.map}{Return mapview leaflet to visualize output of fcn}
}
\description{
Fixes all hwys in place from raw data from subset of raw nhpn data. Note these
functions expect NHPN format, i.e., SIGNT1 and SIGNN1 columns as identifiers.
}
