% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract-lvl-divM-fcns.R
\name{gen.cross.tract.dividedness}
\alias{gen.cross.tract.dividedness}
\title{gen.cross.tract.dividedness}
\usage{
gen.cross.tract.dividedness(
  region,
  divs,
  nbd.query.fcn = tigris::tracts,
  region.id.colm = "rid",
  fill.nhpn.gaps = F,
  erase.water = F,
  year = 2019,
  ...
)
}
\arguments{
\item{nbd.query.fcn}{i.e., tracts or block_groups; passed onto
\code{geox::tracts.from.sf}}

\item{region.id.colm}{identifer column for supplied region}

\item{...}{passed onto \code{polygonal.div}}

\item{cutout.water}{wheter or not to cut out water areas before getting
dividedness.}
}
\value{
a data.frame with one row per census tract in the supplied region, with a
geoid column and a poly.id column indicating which "polygon division" each tract
is in. Tracts with different identifiers in this column are across a division
from one another.
}
\description{
Given a large region: Place, CZ, CBSA, etc., and a division object, represented by
an \code{sf} object with only linear features, allocate each neighborhood, defined as
tracts or block groups, to a side of the division within the area.
}
