% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{Get.bundled.ray.output}
\alias{Get.bundled.ray.output}
\title{Get.bundled.ray.output}
\usage{
Get.bundled.ray.output(place, trimmed.hwys, include.map = TRUE, ...)
}
\arguments{
\item{trimmed.hwys}{hwys prepped for supplied place, as with
\code{initial.hwy2ray.subset}}

\item{...}{additional parameters passed onto mapview, if map is being returned.}
}
\description{
Calls \code{hwys2endpoints}, which finds all line start/endpoints t whether each
constitutes a ray. This function counts rays up and bundles with a map, if
\code{include.map} is TRUE
}
