% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys fcns.R, R/spatial clean fcns.R
\name{handle.overlaps}
\alias{handle.overlaps}
\title{handle.overlaps}
\usage{
handle.overlaps(x)

handle.overlaps(x)
}
\description{
Some shapes (loops, conjoined loops) get a series of overlapping
polygons from st_split. Example, minneapolis BTS rails. This cuts out areas
so all sub polygons are non-overlapping.

This cuts out areas so all sub polygons are non-overlapping. Trims out
overlaps from larger shapes, while leaving overlapping shape on its own.
Relevant when a smaller polygon is 100\% covered by larger one(s) and you want
non-overlapping polygons.
}
