% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div proximity fcns.R
\name{local_equidistant_project}
\alias{local_equidistant_project}
\title{this script provides functions for getting the proportion of the populations
within x proximity of a division, and the mean distance to division for CT/CBG
centroids; these were initial dividedness measures.}
\usage{
local_equidistant_project(x, projection_center = NULL)
}
\arguments{
\item{x}{sf to project}

\item{projection_center}{to center projection around.}
}
\description{
Projects a shapefile to an Azimuth Equidistant Projection centered on a
supplied centroid, or if \code{projection_center} is null, on a centroid
automatically generated for the supplied sf object.
}
\details{
There's also a set of functions for re-projecting spatial data based on its centroid.
This can be used to use distance- or area-preserving map projections centered on
local areas. This should only be necessary when a function that can only operate
on planar geometries is needed.
local_equidistant_project
}
