% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys fcns.R
\name{polygonal.div}
\alias{polygonal.div}
\title{polygonal.div}
\usage{
polygonal.div(
  region,
  divs,
  negative.buffer = 100,
  min.size = NULL,
  verbose = F,
  return.sf = F,
  ...
)
}
\arguments{
\item{region}{\code{sf} object representing neighborhoods to allocate to subdivision}

\item{divs}{linear \code{sf} representing divisions.}

\item{negative.buffer}{Shrink region by this amount only for calculating polygons.
Useful for handling shpfiles w/ different resolutions, especially for regions
along an international border. Defaults to 100m.}

\item{min.size}{Minimum (area) size for population, in meters. Defaults to 5e5, or
1/2 a km^2. Set to NULL to not filter by size.}

\item{return.sf}{If true, returns an sf object, w/ one row per polygon
subdivision, that can easily be mapped or allocated to neighborhoods. If false
(default), returns number of polygons.}
}
\description{
Given a boundary and corresponding divisions within boundary,  finds number of
sub-polygons in region. Returns a dataframe with region identifiers and number of
polygons. Filter all subpolygons w/ area between 50,000 m^2 (these comprise
highway interchanges for example.) Also applies "negative buffer" -- i.e., shrinks
the area to find polygons. This handles issues where hwy resolution is slightly
different from that of CZs or ends slightly earlier at international boundary. It
also means that if a hwy ends or changes class a few meters from the cz boundary,
it can still count as a division. (Operative difference in Dallas, TX, for
example)
}
