% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract-lvl-divM-fcns.R
\name{tracts.across.division}
\alias{tracts.across.division}
\title{tracts.across.division}
\usage{
tracts.across.division(div, ctsf, cutout.water = F, ...)
}
\arguments{
\item{div}{sf containing only linear features. Can have national features;
subsetted within function}

\item{...}{passed onto \code{polygonal.div}}
}
\description{
Given a subset of CTs and a set of linear division, gets whether the tracts are
across the division from one another
}
