% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db or api helper fcns.R
\name{tracts.from.region}
\alias{tracts.from.region}
\title{tracts.from.region}
\usage{
tracts.from.region(region, query_fcn = tigris::tracts, cutout.water = F, ...)
}
\arguments{
\item{query_fcn}{function to query geos from census api. Defualt gets tracts}

\item{...}{passed onto \code{query_fcn}}

\item{region.ids}{1-row tibble with region id/name/type, as returned by
\code{get.region.identifiers}.}
}
\value{
associated census tracts (or other geometry queried from census api)
}
\description{
tracts.from.region
}
