% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polys fcns.R
\name{trim.polys.by.pop}
\alias{trim.polys.by.pop}
\title{trim.polys.by.pop}
\usage{
trim.polys.by.pop(
  sub.polys,
  region.id,
  region.type = "cz",
  min.population.count = NULL,
  min.population.perc = NULL,
  return.sf
)
}
\description{
Trims polygons by population. Helper fcn called from polygonal.div if min.population has a non-null value.
CT table with population and tract identifiers is hardcoded in; expects 'cts' table in environment.
}
