% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{trim.to.length.floor}
\alias{trim.to.length.floor}
\title{trim.to.length.floor}
\usage{
trim.to.length.floor(
  region,
  divisions,
  length.floor = 1000,
  id.col = "SIGN1",
  verbose = T,
  ...
)
}
\arguments{
\item{length.floor}{Filters highways by route if the total length of all portions
of the highway within the area is below this threshold. Default is 1 km.}

\item{id.col}{specifies what column must have minimum \code{length.floor} within
the region. I.e., it should be SIGN1 when filtering out NHPN data grouped by
this column}

\item{verbose}{Display additional text output in console. Makes explicit some
parameters that are passed to wrapped fcns and will say where ineligible rays
are removed due to issue described above.}
}
\description{
Trims a set of highways to those for which the whole stretch of highway inside the
region has minimum length \code{length.floor} in meters. 1 km by default.
}
