% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places and rays fcns.R
\name{trim.to.length.floors}
\alias{trim.to.length.floors}
\title{trim.to.length.floors}
\usage{
trim.to.length.floors(
  region,
  divisions,
  minimum.segment.length = 10,
  minimum.hwy.length = 1000,
  ...
)
}
\arguments{
\item{minimum.segment.length}{Minimum length (meters) that each highway ~segment~
must have in order to be eligible for rays, This differs from
\code{minimum.hwy.length} in that it counts each separate}

\item{minimum.hwy.length}{Minimum length of each ~highway~ that must be inside
place boundaries for it to be ray-eligible. Differs from minimum.segment.length
in that at least one segment of hwy must meet this threshold for ~any~ of the
segments to be eligible. Only relevant if larger than
minimum.segment.length and}
}
\description{
Trims a set of highways to those for which the whole stretch of highway inside the
region has minimum length \code{length.floor} in meters. 1 km by default.
}
