% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract-lvl-divM-fcns.R
\name{Wrapper_gen.tract.within.distance}
\alias{Wrapper_gen.tract.within.distance}
\title{Wrapper_gen.tract.within.distance}
\usage{
Wrapper_gen.tract.within.distance(
  cz = NULL,
  cbsa = NULL,
  divs,
  cutout.water = F,
  clean.nhpn = F,
  ...
)
}
\arguments{
\item{cz}{one of a cz or cbsa identifier code (either 5-digit # or 5-char
numeric)}

\item{cbsa}{one of a cz or cbsa identifier code (either 5-digit # or 5-char
numeric)}

\item{clean.nhpn}{Logical; whether or not to clean divs as if they are
NHPN hwy data by applying cleaning fcns \code{denode.lines} and \code{Fix.all.hwys}. Can
be a vector of length \code{divs} to clean for some measures.}

\item{...}{passed onto \code{tracts.across.division} and/or \code{subset.polys.divs}}
}
\description{
Wrapper_gen.tract.within.distance
}
